requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
google-cloud-aiplatform
pandas